Move sub-classes of "DomainUpDown", "Forms" and "DropDownButton" classes to their own files
