+++ src/System.Windows.Forms/src/System/Windows/Forms/DomainUpDown.cs
@@ -545,245 +545,5 @@ namespace System.Windows.Forms
             width = SizeFromClientSize(width, height).Width + _upDownButtons.Width;
             return new Size(width, height) + Padding.Size;
         }
-
-        // DomainUpDown collection class
-
-        /// <summary>
-        ///  Encapsulates a collection of objects for use by the <see cref='DomainUpDown'/>
-        ///  class.
-        /// </summary>
-        public class DomainUpDownItemCollection : ArrayList
-        {
-            readonly DomainUpDown owner;
-
-            internal DomainUpDownItemCollection(DomainUpDown owner)
-            : base()
-            {
-                this.owner = owner;
-            }
-
-            [Browsable(false)]
-            [DesignerSerializationVisibility(DesignerSerializationVisibility.Hidden)]
-            public override object this[int index]
-            {
-                get
-                {
-                    return base[index];
-                }
-
-                set
-                {
-                    base[index] = value;
-
-                    if (owner.SelectedIndex == index)
-                    {
-                        owner.SelectIndex(index);
-                    }
-
-                    if (owner.Sorted)
-                    {
-                        owner.SortDomainItems();
-                    }
-                }
-            }
-
-            /// <summary>
-            /// </summary>
-            public override int Add(object item)
-            {
-                // Overridden to perform sorting after adding an item
-
-                int ret = base.Add(item);
-                if (owner.Sorted)
-                {
-                    owner.SortDomainItems();
-                }
-                return ret;
-            }
-
-            /// <summary>
-            /// </summary>
-            public override void Remove(object item)
-            {
-                int index = IndexOf(item);
-
-                if (index == -1)
-                {
-                    throw new ArgumentOutOfRangeException(nameof(item), item, string.Format(SR.InvalidArgument, nameof(item), item));
-                }
-                else
-                {
-                    RemoveAt(index);
-                }
-            }
-
-            /// <summary>
-            /// </summary>
-            public override void RemoveAt(int item)
-            {
-                // Overridden to update the domain index if necessary
-                base.RemoveAt(item);
-
-                if (item < owner._domainIndex)
-                {
-                    // The item removed was before the currently selected item
-                    owner.SelectIndex(owner._domainIndex - 1);
-                }
-                else if (item == owner._domainIndex)
-                {
-                    // The currently selected item was removed
-                    //
-                    owner.SelectIndex(-1);
-                }
-            }
-
-            /// <summary>
-            /// </summary>
-            public override void Insert(int index, object item)
-            {
-                base.Insert(index, item);
-                if (owner.Sorted)
-                {
-                    owner.SortDomainItems();
-                }
-            }
-        } // end class DomainUpDownItemCollection
-
-        private sealed class DomainUpDownItemCompare : IComparer
-        {
-            public int Compare(object p, object q)
-            {
-                if (p == q)
-                {
-                    return 0;
-                }
-
-                if (p is null || q is null)
-                {
-                    return 0;
-                }
-
-                return string.Compare(p.ToString(), q.ToString(), false, CultureInfo.CurrentCulture);
-            }
-        }
-
-        internal class DomainItemListAccessibleObject : AccessibleObject
-        {
-            private readonly DomainUpDownAccessibleObject parent;
-
-            public DomainItemListAccessibleObject(DomainUpDownAccessibleObject parent) : base()
-            {
-                this.parent = parent;
-            }
-
-            public override string Name
-            {
-                get
-                {
-                    string baseName = base.Name;
-                    if (baseName is null || baseName.Length == 0)
-                    {
-                        return "Items";
-                    }
-                    return baseName;
-                }
-                set => base.Name = value;
-            }
-
-            public override AccessibleObject Parent
-            {
-                get
-                {
-                    return parent;
-                }
-            }
-
-            public override AccessibleRole Role
-            {
-                get
-                {
-                    return AccessibleRole.List;
-                }
-            }
-
-            public override AccessibleStates State
-            {
-                get
-                {
-                    return AccessibleStates.Invisible | AccessibleStates.Offscreen;
-                }
-            }
-
-            public override AccessibleObject GetChild(int index)
-            {
-                if (index >= 0 && index < GetChildCount())
-                {
-                    return new DomainItemAccessibleObject(((DomainUpDown)parent.Owner).Items[index].ToString(), this);
-                }
-
-                return null;
-            }
-
-            public override int GetChildCount()
-            {
-                return ((DomainUpDown)parent.Owner).Items.Count;
-            }
-        }
-
-        public class DomainItemAccessibleObject : AccessibleObject
-        {
-            private string name;
-            private readonly DomainItemListAccessibleObject parent;
-
-            public DomainItemAccessibleObject(string name, AccessibleObject parent) : base()
-            {
-                this.name = name;
-                this.parent = (DomainItemListAccessibleObject)parent;
-            }
-
-            public override string Name
-            {
-                get
-                {
-                    return name;
-                }
-                set
-                {
-                    name = value;
-                }
-            }
-
-            public override AccessibleObject Parent
-            {
-                get
-                {
-                    return parent;
-                }
-            }
-
-            public override AccessibleRole Role
-            {
-                get
-                {
-                    return AccessibleRole.ListItem;
-                }
-            }
-
-            public override AccessibleStates State
-            {
-                get
-                {
-                    return AccessibleStates.Selectable;
-                }
-            }
-
-            public override string Value
-            {
-                get
-                {
-                    return name;
-                }
-            }
-        }
     }
 }
